pyproject.toml
[project]
name = "n-mars"
version = "0.1.0"
description = "Non-Monotonic Autoregressive Sequence Models"
requires-python = ">=3.11,<3.13"
dependencies = [
    "ms-swift[llm]",
    "lm-eval",
    "transformers",
    "datasets",
    "wandb",
    "pyyaml",
    "torch",
    "accelerate",
    "peft",
    "trl",
    "flash-attn",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/n_mars", "src/baselines"]

[tool.uv]
dev-dependencies = [
    "ruff",
    "pytest",
    "ipython",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[tool.uv.pip]
no-build-isolation-package = ["flash-attn"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
